Finish ICI subpackage, add datasets -> still need tests
