+++ src/ampworks/__init__.py
@@ -12,46 +12,56 @@ includes search functionality and more detailed examples.
 
 """
 
+from ._core import (
+    Dataset,
+    read_csv,
+    read_excel,
+    read_table,
+)
+
 __version__ = '0.0.2.dev0'
 
 __all__ = [
-    'dqdv',
-    'gitt',
+    'Dataset',
+    'read_csv',
+    'read_excel',
+    'read_table',
     'ici',
-    'data',
-    'plotutils',
+    'gitt',
+    'dqdv',
     'utils',
+    'datasets',
+    'plotutils',
+    '_in_interactive',
+    '_in_notebook',
 ]
 
 
-def __getattr__(attr):
+# Lazily load submodules
+_lazy_modules = {
+    'ici': 'ampworks.ici',
+    'gitt': 'ampworks.gitt',
+    'dqdv': 'ampworks.dqdv',
+    'utils': 'ampworks.utils',
+    'datasets': 'ampworks.datasets',
+    'plotutils': 'ampworks.plotutils',
+}
 
-    if attr == 'dqdv':
-        import ampworks.dqdv as dqdv
-        return dqdv
-    elif attr == 'gitt':
-        import ampworks.gitt as gitt
-        return gitt
-    elif attr == 'ici':
-        import ampworks.ici as ici
-        return ici
-    elif attr == 'data':
-        import ampworks.data as data
-        return data
-    elif attr == 'plotutils':
-        import ampworks.plotutils as plotutils
-        return plotutils
-    elif attr == 'utils':
-        import ampworks.utils as utils
-        return utils
+
+def __getattr__(name):
+    if name in _lazy_modules:
+        module = __import__(_lazy_modules[name], fromlist=[name])
+        globals()[name] = module  # cache for later
+        return module
+    raise AttributeError(f"module {__name__} has no attribute {name}")
 
 
 def __dir__():
-    public_symbols = (globals().keys() | __all__)
-    return list(public_symbols)
+    return list(globals().keys() | __all__)
 
 
-def _in_interactive():
+# Check for interactive and notebook environments
+def _in_interactive() -> bool:
     try:
         from IPython import get_ipython
         return get_ipython() is not None
@@ -59,7 +69,7 @@ def _in_interactive():
         return False
 
 
-def _in_notebook():
+def _in_notebook() -> bool:
     try:
         from IPython import get_ipython
         shell = get_ipython().__class__.__name__
