--- src/ampworks/data/__init__.py
@@ -1,13 +1,0 @@
-"""
-TODO
-"""
-
-from ._dataset import Dataset
-from ._read import read_csv, read_excel, read_table
-
-__all__ = [
-    'Dataset',
-    'read_csv',
-    'read_excel',
-    'read_table',
-]
